Sentiment Analysis/requirements.txt
joblib==1.0.1
nltk==3.6.7
numpy==1.21.0
pandas==1.3.0
Pillow==9.0.1
scikit_learn==1.0.2
streamlit==1.7.0
